+++ config/indonesia.php
@@ -1,25 +1,55 @@
 <?php
 
+/**
+ * Read more at https://github.com/kodepandai/laravel-indonesia.
+ */
 return [
     /**
-     * TODO: docs
+     * Table prefix for indonesia tables: provinces, cities, districts and villages.
      */
     'table_prefix' => 'indonesia_',
 
     /**
-     * TODO: docs
+     * API Configuration.
      */
     'api' => [
+        /**
+         * If enabled, this will load Indonesia API.
+         * - http://localhost:8000/api/indonesia/provinces
+         * - http://localhost:8000/api/indonesia/cities
+         * - http://localhost:8000/api/indonesia/districts
+         * - http://localhost:8000/api/indonesia/villages
+         */
         'enabled' => true,
 
+        /**
+         * The middleware for Indonesia API.
+         */
         'middleware' => ['api'],
 
+        /**
+         * The route name for Indonesia API.
+         */
         'route_name' => 'api.indonesia',
 
+        /**
+         * The route prefix for Indonesia API.
+         */
         'route_prefix' => 'api/indonesia',
 
-        'response_cached' => true,
+        /**
+         * Specify which column will be displayed in the response data.
+         * Only support columns from database.
+         */
+        'reponse_columns' => [
+            //.
+            'province' => ['code', 'name'],
 
-        'response_include_latitude_longitude' => false,
+            'city' => ['code', 'province_code', 'name'],
+
+            'district' => ['code', 'city_code', 'name'],
+
+            'village' => ['code', 'district_code', 'name'],
+        ],
     ],
 ];