+++ src/IndonesiaServiceProvider.php
@@ -2,7 +2,6 @@
 
 namespace KodePandai\Indonesia;
 
-use Illuminate\Support\Facades\Config;
 use Illuminate\Support\ServiceProvider;
 
 class IndonesiaServiceProvider extends ServiceProvider