+++ CHANGELOG.md
@@ -4,4 +4,11 @@ All notable changes to `laravel-indonesia` will be documented in this file.
 
 ## v1.0.0 
 
+Rework from [laravol/indonesia](https://github.com/laravel/indonesia)
+
 ### Added
+
+- Configuration for database table and API settings
+- API for provinces, cities, districts and villages
+- Database raw data in gzip format (smaller size than csv)
+- Database seeder class to seed Indonesia data